Use latest go-flags
Updated in order to get the "required minimum count for
   positional arguments" functionality
